forensiclens_backend (1)/app/ai_engine/anomaly_detector.py
"""Anomaly Detection AI Engine

This module provides the interface for anomaly detection in forensic photos.
The AI team should implement the actual model inference here.
"""
import torch
import numpy as np
from typing import Dict, List, Tuple, Optional
from PIL import Image
import logging
from app.core.config import settings

logger = logging.getLogger(__name__)


class AnomalyDetector:
    """Anomaly detection model interface
    
    This class should be implemented by the AI team to integrate their
    anomaly detection model. The interface is designed to work with GPU
    acceleration on Google Colab.
    
    Expected functionality:
    - Load pre-trained anomaly detection model
    - Detect tampering, manipulations, and inconsistencies in photos
    - Return anomaly score and detailed regions
    """
    
    def __init__(self, model_path: Optional[str] = None):
        """Initialize the anomaly detector
        
        Args:
            model_path: Path to the pre-trained model
        """
        self.model_path = model_path or settings.ANOMALY_MODEL_PATH
        self.device = "cuda" if settings.USE_GPU and torch.cuda.is_available() else "cpu"
        self.model = None
        
        logger.info(f"Initializing AnomalyDetector on {self.device}")
        self._load_model()
    
    def _load_model(self):
        """Load the pre-trained model
        
        TODO: AI team should implement model loading logic here
        Example:
            self.model = torch.load(self.model_path)
            self.model.to(self.device)
            self.model.eval()
        """
        logger.info("Loading anomaly detection model...")
        # Placeholder: AI team implements actual model loading
        logger.warning("Using placeholder anomaly detector. Implement actual model loading.")
        self.model = None  # Replace with actual model
    
    def preprocess_image(self, image: Image.Image) -> torch.Tensor:
        """Preprocess image for model input
        
        Args:
            image: PIL Image object
            
        Returns:
            Preprocessed tensor ready for model inference
        
        TODO: AI team should implement preprocessing pipeline
        """
        # Placeholder preprocessing
        # Replace with actual preprocessing (resize, normalize, etc.)
        image_array = np.array(image)
        tensor = torch.from_numpy(image_array).float()
        tensor = tensor.permute(2, 0, 1).unsqueeze(0)  # [B, C, H, W]
        return tensor.to(self.device)
    
    def detect(self, image_path: str) -> Dict:
        """Detect anomalies in an image
        
        Args:
            image_path: Path to the image file
            
        Returns:
            Dictionary containing:
            - has_anomalies (bool): Whether anomalies were detected
            - anomaly_score (float): Overall anomaly score (0.0 to 1.0)
            - anomaly_details (dict): Detailed information about detected anomalies
            - anomaly_regions (list): List of bounding boxes for anomalous regions
                Each region: {x, y, width, height, confidence, anomaly_type}
        
        TODO: AI team should implement actual inference logic
        """
        try:
            logger.info(f"Detecting anomalies in {image_path}")
            
            # Load image
            image = Image.open(image_path).convert("RGB")
            
            # Preprocess
            input_tensor = self.preprocess_image(image)
            
            # Inference
            # TODO: AI team implements actual model inference
            # with torch.no_grad():
            #     output = self.model(input_tensor)
            #     anomaly_score, regions = self._postprocess(output)
            
            # Placeholder response
            logger.warning("Using placeholder anomaly detection results")
            return {
                "has_anomalies": False,
                "anomaly_score": 0.15,
                "anomaly_details": {
                    "model_version": "placeholder_v1",
                    "detection_types": ["tampering", "splicing", "copy-move"],
                    "confidence": 0.85
                },
                "anomaly_regions": [
                    # Example region format
                    # {
                    #     "x": 100,
                    #     "y": 150,
                    #     "width": 200,
                    #     "height": 200,
                    #     "confidence": 0.9,
                    #     "anomaly_type": "splicing"
                    # }
                ]
            }
        
        except Exception as e:
            logger.error(f"Error in anomaly detection: {str(e)}")
            raise
    
    def batch_detect(self, image_paths: List[str]) -> List[Dict]:
        """Detect anomalies in multiple images (batch processing)
        
        Args:
            image_paths: List of image file paths
            
        Returns:
            List of detection results for each image
        
        TODO: AI team can implement optimized batch processing
        """
        results = []
        for image_path in image_paths:
            result = self.detect(image_path)
            results.append(result)
        return results


# Singleton instance
_anomaly_detector = None


def get_anomaly_detector() -> AnomalyDetector:
    """Get or create the anomaly detector singleton instance"""
    global _anomaly_detector
    if _anomaly_detector is None:
        _anomaly_detector = AnomalyDetector()
    return _anomaly_detector
