forensiclens_backend (1)/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# AI/ML Libraries
torch==2.1.2
torchvision==0.16.2
Pillow==10.2.0
numpy==1.26.3

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# ngrok for Colab deployment
pyngrok==7.0.5

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utilities
python-slugify==8.0.1
